+++ src/NKDRoyalMailBarcode.m
@@ -208,7 +208,7 @@ static float royalmail_barBottom( int hexDigit ) {
     // (10101010) - checksum
     // 1 - terminator
 
-    int contentLength = [[self content] length];
+    NSUInteger contentLength = [[self content] length];
     int hexDigit;
     int descriptor = 0;
